Fix CtsAppSearchTestCases Failed issue.
The GlobalSearch test item uses "body" as the search keyword. The expected search results should only contain the test data from within that test item. In a mobile phone environment, no app matches this condition. However, in an XR environment, the test determination fails because the "Body Tracking Sample App" is included in the search results.

Our solution is to filter the search data to exclude irrelevant content. This ensures that apps on the device do not affect the execution of CTS tests.

Bug: 407607207
Bug: 407607879
Test: local test and device.
    atest CtsAppSearchTestCases
Flag: EXEMPT bugfix
Change-Id: I3e9167d8fa900a72ff4104721be9d4d15fb99f0c