Run testStartUsingNetworkFeature_enableHipri on gsi
Currently, cts-on-gsi does not run testStartUsingNetworkFeature_enableHipri
because this test is in the cts-on-aosp-exclude.xml, which is  included
from cts-on-gsi.xml
This exclusion was done by ag/2532035 in 2017, but the original reason
is no longer known.
There is currently no known issue preventing this test from running on
gsi.
This CL removes the test from exclude.xml to see if there will be an
issue.

Bug: 409706191
Bug: 408900604
Test: TH
Flag: TEST_ONLY
Change-Id: I2f2ee804f57ec603ab7683a9b26e4f0b15c4d387