Allow freeform task to be launched into the launch-root in the test
The test requests to launch an Activity as fullscreen and position
its Task under a multi-window Task, which is a weird use case. In
such a case, the windowingMode is resolved as fullscreen in
TaskLaunchParamsModifier, while the returning windowingMode is
undefined (because resolve mode is the same as the display).
So, the leaf Task is still in multi-window mode (inherited from
the parent Task, i.e. split-secondary). Meanwhile, it also returns
a freeform bounds because both freeform and fullscreen windowingMode
will calculate the freeform bounds (since commit b3968b8). Therefore,
the leaf task is updated to a freeform bounds, which is smaller
than the split-secondary root, and won't be able to occluded the
existing activities on the split-secondary.

This test currently works because the leaf task now always inherits
the bounds of its organized parent Task (see
Task#updateOverrideConfigurationFromLaunchBounds). But this test
would fail when the policy is changed.

Update the test to accommodate the freeform windowingMode to
directly allow activity/task to be launched into the launch-root.

Bug: 410745462
Test: MultiWindowTests on phone and cf_desktop
Flag: EXEMPT test only
Change-Id: I8add5493773f14c5c219e3caf776945b5b3ebf53