Wait screen on before pressing Menu button
During the test_screen_on_only_off_host_service test, the script tries
to press the Menu button before confirming the screen is on.
This can lead to a concurrency issue which causes the
waitAndGet('ScreenOn') to timeout.

Change the order to first wait for the screen to be on, and then press
the Menu button.

Bug: b/411401777

Change-Id: Iff2c54e880f221391bd79bebe290c0fb13550f88
Signed-off-by: Daniel Finimundi <danielrf@motorola.com>