add flag libary in dependency
This test uses flags from library
android.service.autofill.flags-aconfig-java, but it doesn't include this
library as an dependency. At runtime it will load the flag class from
the framework, which is wrong, since this flag is not an API, it is not
guaranteed that the flag will always be on the device, and it is not
guaranteed that the when the test is running, always the expected flag
value will be loaded.

Test: atest CtsAutoFillServiceTestCases
Flag: EXEMPT add dependency
Bug: 411715086
Change-Id: I0fddfb4d1e7faf3819eb61e0798eb242e6b61c10