Offset tap to avoid rounded corners in tests
The root cause of these two failed cases is that the test code tried to
tap the top-left corner, but the window's rounded corners caused the tap
to fail.
My suggestion is that add offset to make it tap the bottom-right position to avoid the top-left rounded corner.

Cherry pick from: https://android-review.git.corp.google.com/c/platform/cts/+/3603372

Bug: 349160052
Bug: 467623640
Test: atest android.view.surfacecontrol.cts.SurfaceControlViewHostTests with the device and no issues observed.
Change-Id: I2976ffb64b504aabaff8ae20a8124c34f76adc80