input: replace EventVerifier with BlockingQueueEventVerifier
The two classes are almost identical, and every use of EventVerifier is
actually backed by a blocking queue anyway, so let's just have one
class. In most cases, this change is as simple as using the
BlockingQueueEventVerifier created by CaptureEventActivity, or passing
the queue directly to the BlockingQueueEventVerifier constructor rather
than a lambda that polls it. The exception is
UinputRecordingIntegrationTests, where we need to add a new mode to
CaptureEventActivity in which batched events are split before being
added to the queue.

Bug: 245989146
Test: $ atest CtsInputTestCases CtsInputHostTestCases \
        CtsSecurityTestCases:android.security.cts.FlagSlipperyTest \
        InputTests:com.android.test.input.UinputRecordingIntegrationTests
Test: $ m checkinput
Flag: TEST_ONLY
Change-Id: I63dbfdbdebd9d7eec4c65386fe4fd3deaa4df65c