AudioPlaybackConfigurationTest: use test ID to catch APC updates
Use a test ID in the AudioAttributes of the players used for
testing AudioPlaybackConfiguration so they can be matched to
the configuration updates in the callback, to more reliably
identify changes to the state of a player that are tracked
in the plabyack callback.
The matching is done by replacing the counting of the callbacks
(which is not accurate anymore after optimizations on the
config callback) by the evaluation of a Predicate. This
predicate checks that the configs in the callback contain
AudioAttributes with the injected test ID.

Bug: 410700323
Flag: EXEMPT bug fix
Test: atest android.media.audio.cts.AudioPlaybackConfigurationTest
Change-Id: I31158527a8be6f5e67ee44a77fd96b5bac561eef