Verify outgoing call is dialing before setting active
This improves the CUJ tests to assert that outgoing calls are in the dialing state before they transition to active.  This makes them more robust and follows the correct sequence of states for a call. There was flakiness observed in some test runs where the call moves back to dialing after being set to active.

Bug: 414120755
Flag: EXEMPT test fix
Test: atest UnholdableCallTest
Change-Id: I3052a1edbf1abc3d72d966994fef8de805ab7457