remove flags check in test
This change removes checking of the following flags

1. FLAG_CERTIFICATE_TRANSPARENCY_CONFIGURATION
2. FLAG_CERTIFICATE_TRANSPARENCY_PLATFORM

The reason is in the test set up the these CTS tests haven't set up
dependencies to networksecurity-aconfig-flags-lib. So at runtime these
CTS tests will load the library from conscrypt library on the device.
This is not expected behavior, since for the released CTS, the CTS
should load the flag value from the combiled CTS binary, but not from
the devices.

Test: atest CtsNetSecConfigCertificateTransparencyTestCases
NetSecConfigCertificateTransparencySctLogListTestCases
CtsNetSecConfigCertificateTransparencyDefaultTestCases
NetSecConfigCertificateTransparencySecNoLogListTestCases
Bug: 303773055
Flag: EXEMPT remove flags

Change-Id: I089ded113a70c7f8e98e1227e6ece8778fa5ade2