Ignore UnknownHostException in TestUtils
When establishing a connection, even if the internet is supposedly
available, an error might be raised when trying to resolve the hostname.
In this case, this is not a failure of the test, but a failure of the
assumptions (that the network is usable).

While there, move from the deprecated junit.framework package to
org.junit. Remove the subclassing to Assert.

Bug: 415997711
Test: atest CtsNetSecConfigBasicDomainConfigTestCases
Flag: EXEMPT bugfix
Change-Id: If7f7ca54a48ba9acae471ceb8f3ed61d7634c967