CTS Verifier: use a FIFO instead of reverse JNI
The Audio Data Path and some other tests were using reverse
JNI in a callback. That can block and cause glitches.

So it was replaced with an atomic FIFO that can be written
by C++ and read by Java.

This change required removing the JavaSinkHandler,
which was also causing data corruption.

Also fix some memory leaks.

Bug: 376133542
Bug: 418303038
Test: CTS Verifier Audio Data Path tests
Test: and Tap-to-Tone
Flag: TEST_ONLY
Change-Id: I2d799eac29e505c138c4381bdf3593eb8bf73ec4