Restoring ported DisplayEventTest back to cts suite
The DisplayEventTest were ported from cts to unit tests because of a permission issue with testing the frozen process feature. While havig the frozen process tests in unit tests is fine, we should continue to have the tests for DisplayEventTest in cts. This is essentially restoring ag/29597564

Bug: 417191815
Test: atest android.display.cts.DisplayEventDeliveryTest
Flag: EXEMPT test only
Change-Id: I6e97e96379ff773453e0eacf058964998826112e