Query audio patch to identify 24 bit support.
When opening low latency stream on mmap path, pcm float is used for
opening the stream. If it is not supported, APM may suggest a pcm 16 bit
even if greater than 24bit is supported. In that case, query the
supported formats for the mix port used for the audio patch can be a fallback logic to get a more reliable test result.

Bug: 407643937
Test: run the test
Flag: TEST_ONLY
Change-Id: I5c049e28622401d101e29eed48ab88306b082ea4