Attempt to fix "process crashed" during test
When the app is reinstalled, system kills all existing processes. This
happens asynchronously on system_server handler thread. Sometimes it
happens when running subsequent tests and leads to failure. This CL
makes the test wait for a message in logcat that is emitted when the
system kills the processes.

Bug: 408089011
Test: atest CtsDevicePolicyManagerTestCases:com.android.cts.devicepolicy.DeviceOwnerTest#testListForegroundAffiliatedUsers_extraUser
Flag: EXEMPT test change
Change-Id: I93389f7b56bbf3626f285869eef6561f2ce17acf