Make IME Switcher Menu tests ID agnostic
The IME Switcher Menu tests were using specific IDs to target views,
facilitating the testing flow. However, there is no strong requirement
on the UI being implemented a specific way, or having specific IDs.

This modifies the tests to not rely on IDs, but still allows for finding
a scrollable view and scrolling into it, to accomodate small screen
devices.

This also moves the tests related to the IME Switcher Menu Settings
button out of CTS, as this was not required to be implemented as part of
the UI.

Flag: EXEMPT testfix
Bug: 416313466
Test: atest InputMethodManagerTest#testInputMethodPickerShownItems
  InputMethodManagerTest#testInputMethodPickerSwitchIme
Change-Id: I9c526fbc614a5c5b46839edbe54044039fc4556e