[AOA FFS Implementation] Add Test for HID Protocols
This change add tests for the HID Protocols in AOA. It tests that device
is registered and is able to send events. Also verifies that expected
events are processed at the device side. It tries to test in following
modes.
- Send HID Descriptor in Full
- Send HID Descriptor in Parts
- Send multiple HID Descriptors interleaved
It also test that if device is not registered then sending
descriptors/events are not processed.
- Send Descriptor to unregistered device.
- Send Event to unregistered device.
This will test both flagged and unflagged changes.

MUST_SLEEP

BYPASS_LARGE_CHANGE_WARNING: Adding extensive testing to previously
untested functionality.

Bug: 407985367
Flag: EXEMPT TEST_ONLY
Test: Tested manually that all the tests are passing.
Change-Id: I560a578b80c625a0b13cd4a13919210133cf397f