media: signal eos if no further inputs will be queued
HDRDecoderTest and HdrToSdrDecoderTest are not signalling eos after
final enqueue input call. For components that have latency, they may
hold back final few outputs until eos is seen. As eos is not signalled,
all frames are not sent by the component and the test timesout waiting.
This change addresses this.

Bug: 414559404
Test: atest CtsMediaDecoderTestCases:HDRDecoderTest
Test: atest CtsMediaDecoderTestCases:HdrToSdrDecoderTest
Flag: TEST_ONLY
Change-Id: I010c2147f9c31299468adb3c734bc544357e0231