SystemMediaRouter2Test: reorder init in setup
`setupAndGetService` registers a MediaRouter2 instance internally and
empirically we find that if an external MediaRouter2 is registered
before this on the same context, the internal instance has a high chance
of ending up in removing routes and result in test failures.

This also makes the setup and tear down order between the external
router and service symmetric.

Bug: 419589719
Flag: EXEMPT test only
Test: atest CtsMediaRouterTestCases
Change-Id: I4a75f73361034dbc21836bb3afdd7b1572bab7c4