fix timing issue with call-style notif posts in FGS tests
In b/419590670, the logs showed it was possible for the notification
to reach the VoipCallMonitor class BEFORE the call was added to the
class which is problematic.  Clients should be posting call-style
notifications after adding the call to telecom via TM#addCall.

In order to prevent this timing issue, the initial call-style post
should be moved to when the onResult is delivered to the client. This
guarantees that the call has reached the VoipCallMonitor and eliminates
any timing issues.

Flag: EXEMPT test fix
Fixes: 419590670
Test: atest android.telecom.cts.cuj.app.integration.CallStyleNotificationsTest
Change-Id: Iddefccc6d879190cc2ccea0ce82628e40d911b8e