Add test for registering IME callback ordering
This adds a test to verify that the IME callback is registered and
unregistered on the right dispatchers.

The problem scenario is when going from an Activity A in Process 1, that
requested the IME, to an Activity B in a different Process 2, that
doesn't request the IME, and then to an Activity C in Process 1, that
also doesn't request the IME. In this case, when going from A to B, we
would unregister the callback on A, and register it again on A, instead
of registering on B. If we then go back to a different activity from the
initial process, we would carry over the IME callback that was
re-registered, and capture the next back event instead of allowing the
app to handle it.

Flag: EXEMPT bugfix
Test: atest KeyboardVisibilityControlTest#testHideImeAfterBackPressed_inSecondActivityAfterProcessSwitch
Bug: 407224281
Change-Id: I782113d966a576d216b7b27314ee181ad998a0e5