TouchModeTest: touch the centre of the test activity
Previously the test was touching the top-left of the activity, which was
causing the notification shade to be pulled down. This could break any
subsequent tests that depended on getting a window focused or on top.
While the shade shouldn't show up from touching the top-left of the
activity, touching the centre instead fixes it and should make the test
more robust against similar screen-edge-based disruption in future.

Bug: 420919325
Test: $ atest --no-group-test \
        CtsInputTestCases:android.input.cts.TouchModeTest \
	CtsInputTestCases:android.input.cts.TouchpadAbsoluteCaptureModeTest
Test: $ atest CtsInputTestCases
Flag: TEST_ONLY
Change-Id: I6952731707643d48a195110c610c8321230e3a5c