Test the known cache modules
This adds a CTS test that verifies two things:
 1. A cache can be created with a known module
 2. The cache cannot be invalidated

The second check verifies that that the property behind the cache
cannot be written by the test process.  This stands in for a more
general check that the property behind the cache cannot be written by
any process except the one specifically permitted.

Flag: android.os.ipc_data_cache_module_adservices
Bug: 415306020
Test: atest
 * FrameworksCoreTests:PropertyInvalidatedCacheTests
 * FrameworksCoreTests:IpcDataCacheTest
 * CtsOsTestCases:IpcDataCacheTest

Change-Id: I52d2c9188681c38d3201a31e9aea44d1f3955ab5