Avoid dependency on route ids' shape
Ids are supposed to be opaque objects (with the vague exception of the
default route id). Since they are not part of the public API, we should
not use them in CTS.

This change also removes the direct access to an aconfig flag. This
access was causing some tests to fail.

Bug: b/418636408
Bug: b/419589719
Test: atest CtsMediaHostTestCases
Flag: EXEMPT Test only
Change-Id: Ife2c25e65b632c50c7b4a0b759e47e9da383093b