Only collect StorageIoInterface for Android 15+
Older versions of Android are not assured to have the "/tmp"
directory which we put our script into.  Furthermore, our
analysis of eMMC vs. UFS usage is really only concerns with
current and future devices, not older ones.

So now we don't collect anything if the first API on
the device predates the Android 15 release.

Bug: 422568092
Bug: 422093010
Test: "atest android.edi.cts.StorageIoInterfaceDeviceInfo" passed on device before Android 15, collecting nothing; passed on device after Android 15, collecting our info.
Flag: EXEMPT for bug fix
Change-Id: Ie7a2a1864deb05e846d1246ab9706ba4dc6d78fd