fix CtsWindowManagerDeviceTestCases test fail in TVs
Description: During testing, Launcherx's DispatchActivity is launched,
which is expected to prevent the TurnScreenOnActivity from waking the device.
This results in test failure.

Solution: Adjust the timing for putting the device into sleep mode in the test case,
so that TurnScreenOnActivity can successfully wake up the device and the test can pass..

Bug: 412545355
Flag: EXEMPT cts
Test: run cts -m CtsWindowManagerDeviceTestCases

Signed-off-by: ezra.hu <ezra.hu@mediatek.com>
Change-Id: Iae8ef3a2f8986c3cd8f59e919a17ce0d8ec29f92