create a jarjar rule for cts test
This change creates a jarjar rule for the cts tests. This jarjar rule
renames package com.android.providers.media.flags to
com.android.providers.media.cts.flags.

The reason these flags are used in the photopicker framework, and they are
hidden apis in the framework. CTS test should not access the hidden apis
in the framework. By renaming the flag name, at the runtime CTS will
load the flag value from the tests instead of loading the flag from the
framework. The flag class in photopicker has been renamed, but since the
prebuilt is used in NEXT, the fix in not in the prebuilt

Test: m CtsPhotoPickerTest
Bug: 417904151
Flag: EXEMPT build file change
Change-Id: Ic885e3247def4c820058aa6874ed72899b1c6a02