Allow granting BAL privileged by opt-in on Broadcast PendingIntent
The current logic allows a Broadcast PendingIntent to pass BAL privileges to the broadcast receiver if the PendingIntent has a different creator than the sender and the sender is visible.
The different creator condition makes it impossible for an app to simply create a PendingIntent and send it to pass permissions to the receiver, which is fine when the system is invoking a callback. THe new logic makes this more explicit so that the sender can explicitly opt in.
In the future this should probably the only way to grant BAL privileges via broadcast. A open question is if the API to opt-in on the brpoadcast should be public to allow 3P apps to use this mechanism. This only considers visibility of the app, not other BAL exemptions.

Test: atest BackgroundActivityLaunchTest
Flag: EXEMPT bugfix (maybe?)
Bug: 421197489, 422788436
Change-Id: I6f172d36610f331f64741f6ae75b3800f8a90093