Add support for masking bits in hid output report checks.
Starting with kernel 6.2, DualShock 4 support was moved from the
hid-sony driver to the hid-playstation driver, which has resulted in
some subtle behavior changes.

One change with the hid-playstation driver is to only send device
control commands (rumble, polling interval, etc.) from the kernel driver
to the physical gamepad device when needed. The old driver would send
these with every output report.

These patches introduce a bitmask to the hid output report validation
logic. In order for the tests to function correctly on both the hid-sony
and hid-playstation drivers (as the 6.1 kernel is still supported by
android), we need to mask bits in the output report that will be
different between the two drivers. For instance, updating the polling
rate or setting the rumble motors.

This fixes the DualShock 4 vibration tests.

Bug: 243570601
Test: atest CtsHardwareTestCases
Flag: EXEMPT TEST_ONLY

Change-Id: Ib4c4b8c5eaf3c4a7ea902e14af2ddbd13aac13bb