HDMI-CEC: Improve keycode setting log capture
KEYCODE_SETTINGS might have two behaviors on different devices. If the
overlay is set to 0, which is the default value in PhoneWindowManager,
it would bring up the settings page; if the overlay is set to 1, which
is the behavior most GTV preferred, it would bring up the notification
panel. This patch accounts for both situation. Additionally, log line
capture is updated in order to avoid race condition created by
waitForLog function.

Bug: b/416193425
Test: tested with modified test case and verified by vendor
Flag: EXEMPT test only
Change-Id: Ia4ed761f548e4cc729d39f38e57fca93dc9c2d05