[Autofill CTS] Define an temporary internal annotation to mark certain tests as critical
Engprod team has plan to define a new annotation. The new annotation will be used as a factor to select critical tests. Once the engprod annotation is created, this annotation will be replaced with the new
one.

This annotation will only retain in source code. It's a no-op annotation and should only be used by autofill team to track the progress of annotation critical CTS tests.

This change demonstrates the usage of the annotation by annotating critical test in AugmentedLoginActivityTest

Bug:b/423694164
Flag:EXEMPT Test only
Change-Id: I02890fa40e102fd8a6cb227fa0561789c7c6695d