mediapc: add a precondition to verify multi-instance avc support
When running CodecInitializationTest and FrameDropTest, there is an
instance of mediarecorder and  transcoder is running in parallel.
This requires the device to be capable of running a few AVC
encoders and decoders in parallel. Some low end devices are not
capable of running sufficient number of avc decoder and encoders in
parallel to meet this use-case and hence these tests had a
precondition check to ensure that these are run only if the device
was capable enough to run these instances in parallel.

aosp/3506401 removed this precondtion check and thus resulting in
few low end devices failing these tests. Current change adds the
precondition check in CodecInitializationLatencyTest and
FrameDropTest to handle such cases.

Bug: 412946293
Bug: 423471002
Bug: 422321089
Bug: 419831207
Test: atest android.mediapc.cts.CodecInitializationLatencyTest
Test: atest android.mediapc.cts.FrameDropTest
FLAG: TEST_ONLY

Change-Id: I1435ae987fe374d2c410765bc66702c36973f8b0
Merged-In: I1435ae987fe374d2c410765bc66702c36973f8b0