Fix CTS test CtsMediaMiscTestCases:android.media.misc.cts.NativeImageReaderTest
The failing test was rotating the underlying native window 270 degrees,
expecting the images to report the same rotation. The issue is that the
surface rotation does not affect the image rotation, the camera rotation
does, and there is no guarantee about the camera rotation when running
the tests. The CTS test was modified to correctly cover this.

Test: CTS test passing.
Bug: 419928335
Flag: TEST_ONLY
Change-Id: If97582c61e1460fc6086f82a8af2828e3af263af