Skip testSetDisplayState for automotive
This change excludes the test testSetDisplayState for automotive.

The test case is changing the display states and querying screen on/off
information via PowerManager::isInteractive(). But, in autmotive, even
if the driver display is off, the passenger displays and head up
displays might still remain on because of separate display groups. This
results in isInteractive always returning true and causing the test to
fail

Bug: 425760759
Change-Id: I59b79bf732742df1c85f9795b1adc30c2bafad1a
Signed-off-by: Oleksandr Nikitin <oleksandr.nikitin@bmw.de>