testRepeatingHeic skip a go device
The "GO" platform does not support the camera continuous shot, but this
test directly calls the continuous shot through the low level API. This
will cause the camera process`s memory usage to peak too high, leading
to the lmkd killing the test process and causing the test to fail.
Therefore, it is necessary to skip this test on the "GO" platform.

Flag: EXEMPT bugfix
BUG: 426337110
Test: run cts -m CtsCameraTestCases -t
android.hardware.camera2.cts.ImageReaderTest#testRepeatingHeic[1]

Change-Id: I5e6df198b4a27fedaa4982a9191a6aa47fb25af7