Require FEATURE_DEVICE_ADMIN in wipeData_notAuthorized_throwsException
Current implementation of wipeDataWithReason in
DevicePolicyManagerService simply returns if FEATURE_DEVICE_ADMIN is
disabled. That fails the test, since the test expects the exception in
case of missing permissions.

Require the feature to be enabled in the test, otherwise skip.

Test: atest
android.devicepolicy.cts.WipeDataTest#wipeData_notAuthorized_throwsException
-v

Flag: TEST_ONLY
Bug: b/417959770
Change-Id: Ib3268a11bf63ffc6b31fb392c56840a2f5470aa1