Resolve compat perf regression via @NoLogging annotation
This annotation allows marking specific compat changes so that they are not logged by the compat framework. This is expected to be used in code that is called in performance critical paths.

Note this change also means that non-loggable changes (including targeting older sdk) are no longer going to stats log.

Flag: EXEMPT bugfix
Bug: 424051595
Change-Id: I7c55772418adc6a14c11b028e22937814e4e8b26