Extend AppExitInfo test timeout
Tests are timing out waiting for a process to be started, but it
appears the root cause is that the system has become somewhat slower,
and not that anything is actually broken.  This change doubles the
time the test will wait for a process to be started.

This change also adds the elapsed timer information in an existing
logcat; the extra information is the amount of time the test waited
for an event and the timeout limit.

Flag: EXEMPT test-only
Bug: 426072686
Test: atest
 * CtsAppExitTestCases

Change-Id: I7d95a701ea06dbf3eb33f97ce7d16e00997b84a9