Skip WifiLocationInfoBackgroundTest for automotive
Automotive devices can have additional displays that are configured in
their own display groups. For example:

- Head up Displays
- Passenger displays

Because of this, even if the driver display is off, those displays can
still remain on. This results in the system not reaching a sleep state
and thus the top app in the driver display will not go to a background
state.
Becuase of this, the method 'turnScreenOff()' performed in the
setUpClass never achieves the expected behavior.

Bug: 425760759
Change-Id: Id8ea34899562c450a71a831366e5458494099b26
Signed-off-by: Oleksandr Nikitin <oleksandr.nikitin@bmw.de>