Skip MultidisplayTests for automotive
Automotive devices can have additional displays that are configured in
their own display groups. Because of this, even if the driver display
is off, those displays can still remain on. This results in the system
not reaching a sleep state and thus the top app in the driver display
will not go to a background state.
Becuase of this, the method 'turnScreenOff()' performed in the
setUpClass never achieves the expected behavior.

Bug: 425760759

Change-Id: I2310a7de3c9065168dc8c9a1cf58e695410907a5
Signed-off-by: Oleksandr Nikitin <oleksandr.nikitin@bmw.de>