Make force stop check more precise.
Current code accepts any log message that contains
`Force stopping <package name>`. However it is common that the package
is force-stopped multiple times, e.g. first because it was removed by
the teardown of the previous test and then installed by the current
test. Also, it may be force-stopped in a different user. This CL changes
the check to use a more precise regex.

Bug: 408089011
Test: atest CtsDevicePolicyManagerTestCases:com.android.cts.devicepolicy.DeviceOwnerTest#testListForegroundAffiliatedUsers_extraUser
Flag: TEST_ONLY
Change-Id: I3a3f16331e67ce123dcfec867c032427efffab27