pass the holdability of CS backed calls via extras
A test flake bug came in where the test expects a ConnectionService
backed call to NOT be holdable but upon assertion, the call is holdable.

Upon inspection, this is a race condition where the hold capabilities
have not been cleared yet (since by default all CS backed cals are set
as holdable) and the assertion is made.

To mitigate, the holdability of a call is now conditionally set when
the Connection is created in the ConnectionService.

Flag: EXEMPT test infra update
Bug: 415012173
Test: atest CtsTelecomCujTestCases
Change-Id: I640274ae6cd292fc089dd63bfe3c83fc0df9ad70