fix(Cts): Fix flakiness with SpeakingAccessibilityService init
Root Cause: The test's flakiness stems from a timing dependency,
particularly during periods of high broadcast traffic. The test needs
to wait for the `PACKAGE_ADDED` broadcast to ensure new accessibility
services are fully initialized. If the test proceeds before these
services are ready, it consistently fails with a "not found" error.
Conversely, if it runs after the services are prepared, it passes
without issue. This issue pattern is similar to the issue described
in b/261217661#comment17.

Solution: Based on discussions and insights from b/261217661, the
solution is to use `TestUtils.waitUntil` to ensure the system is
stable before proceeding with the test. Specifically, we will wait
for the test service to be recognized by the AccessibilityManager
before enabling the service.

Bug: 423107875
Test: atest CtsAccessibilityTestCases  --instant
Test: Use go/abtd to perform an A/B test for validate approach
Flag: EXEMPT bugfix
Change-Id: I6efad70969ce4a9f431f44fb04b8f4958050a520