Print sysfs nodes of lights
This will help us debug flakes when the lights are not found.

At the same time, do a slight refactoring of the way we are waiting for
the light nodes: only query for the InputDevice once, and use the same
InputDevice object throughout the rest of the code. Previously, we would
potentially end up with a different (or null) InputDevice object, since
we would query it twice. The returned object thus could have been
different from the one that was originally checked. It's unlikely that
it was causing the failure that we are observing, though.

Bug: 413628576
Test: atest SonyDualshock4BluetoothTest#testLights
Flag: TEST_ONLY
Change-Id: I900bcf2de8001c188dfc7eca8a07f24d94841e03