Dismiss telecom error dialogs for CUJ module
The CtsTelecomCujTestCases module was failing to run all the
way through. Atest would crash around ~40 tests in every time.

Upon inspection, a dialog was popping up around the same time
the test was crashing.  When I removed the tests that presented
the eror dialogs, the module would successfully finish.

This change adds a UiAutomator helper to find the "OK" button
on the dialog and click it, allowing the test to complete.

Flag: EXEMPT test only fix
Fixes: 427797392
Test: atest CtsTelecomCujTestCases
Change-Id: Iaaf978516588a7218c20fabd4d078887edf6ac65