[6/n] Label AutofillCriticalInternal test
For CTS test cases that covers critical user scenario, label it with
AutofillCriticalInternal. This will help us to identify the test cases
that need to be included in the presubmit tests and ensure they are not
flaky.

This is a internal work, it should not impact the current test behavior.

Flag:EXEMPT Test only
Change-Id: I8aae18350e4e08982e3a45bf5bed3311b6249bb1