Rework permission checks in CtsSupervisionTestCases
When `callWithShellPermissionIdentity` is called, all previously
adopted permissions are cleared for the remainder of the test. A new
helper method was added to restore previously adopted permissions.

This was needed to support the change to `setSupervisionEnabled`.

Removed `multiuser` test configs as it shouldn't be needed.

Bug: 395630828
Flag: EXEMPT test only change
Test: atest CtsSupervisionTestCases
Change-Id: Ie2c2f6ec21ecaa5fba4825ea976e00598c668fd8