VQFloor CTS: push resources as per test needs
Before the start of all tests, all resources are pushed to the device.
This is done explicitly by the test and not done via TargetPreparer.
This is causing issues in shard mode. Modify the implementation such
that the resources are pushed by each test as per its needs.

Bug: 430771619
Test: atest CtsVideoQualityFloorHostTest
Flag: TEST_ONLY

Change-Id: Ifb948e59e3bc5908e30436f96a2077e444b1f826