Add tests for checking deeplink presence in preferences
Preferences with SettingsPreferenceMetadata.NO_SENSITIVITY and SettingsPreferenceMetadata.EXPECT_POST_CONFIRMATION can have a deeplink (which needs to be check if it exist), but don't need to.

Preferences with SettingsPreferenceMetadata.DEEPLINK_ONLY should have a non-null deeplink that takes the user to a specific settings screen.

Preferences with SettingsPreferenceMetadata.NO_DIRECT_ACCESS shouldn't have deeplink.

This is a temporary version of the test, checking all available preferences. This will change after implementing filters in infra.

Bug: b/430820203
Flag: TEST_ONLY
Test: atest bedstead-settings-test:PreferenceWithDeepLinkTest

Change-Id: I8af36ea2ef6cd987ecafd7d859b5a84a52d49aea