Add HD+ Call check to CallDetailsTest.
[Description]
As of today, Dialer shows the HD+ icon in the Incall UI only,
when EVS codec is used.

The purpose of this feature is to extend the existing support
to also show the HD+ icon also in the call history.

[Changes]
It was requested to change the CallDetailsTest inside CTS
repo to add the CallLog.Calls.FEATURES_HD_PLUS_CALL feature.

Bug: 428115434
Flag: com.android.server.telecom.flags.hd_plus_call
Test: CTS verified
Change-Id: I7f5b9afbfef665d00dfecee7216f7941fc445bd2