Increase timeouts in virtual audio permission test
Cuttlefish devices can be slow and they need more time to start activities and return a result (which is used in this test). Slightly increase the waiting timeouts for initializing the audio devices and returning a result from the started test activity, this greatly improves cf flakiness.

Bug: 430061638
Test: atest CtsVirtualDevicesAudioTestCases
Flag: TEST_ONLY
Change-Id: I61944f52f5fb15cee987c45d42ad642e2ef2b3a4