Add check for whether the device's vendor version is not lower than Android V
Wi-Fi7 disabled in the Framework is only permitted when the supplicant AIDL version is at least V3.
Since supplicant AIDL V3 corresponds to vendor version Android V,
a check needs to be added to ensure that the disableWifi7 test is only allowed when the vendor version is greater than Android V.

An update network with Wi-Fi7 disabled triggers a network disconnect, waitForDisconnection should be added after the network update operation.

Bug: 433444169
Test: run cts-on-gsi -m CtsWifiTestCases -t  android.net.wifi.cts.WifiManagerTest#testEnableWifi7


Change-Id: Ic938d2e8675f24542b636def211554dbaf1eb1bb