Remove turnScreenOnWithoutShowWhenLocked_doesNotTurnOnVirtualDisplay
The cost of its maintenance is higher than its value.

Some devices don't support keyguard; some do but it's not configured;
the WMState utility doesn't seem to work in all cases; this test case
tests a combination of flags, which are not used in practice. Even the
public docs say to use turnScreenOn together with showWhenLocked.

Note that the issue is with the test setup and the WM test utilities,
the framework logic is correct.

Fix: 430394087
Fix: 430491536
Fix: 430056677
Fix: 429776833
Fix: 429932472
Fix: 430575791
Test: presubmit
Flag: EXEMPT test deletion
Change-Id: Ia23ec86b1e4073f106d109cf597320ff16482bf1