Camera: Relax the video perf. checks on virtual devices
Camera video recording tests are very strict when it comes
to frame drops and delays. The expectations are challenging to
meet on real physical devices and even more so on emulated/virtual
devices that rely on the host machine performance for stability.
Since testing on virtual devices can run on a variety of hosts,
there is no way to guarantee consistent performance and test
stability without relaxing some of the stricter performance
thresholds.

Bug: 427253107
Test: Camera CTS
Change-Id: Id4654452c4589afbcd97d417038e0a30fd05506f