Use UinputTouchScreen in MultiDisplayImeTests
This changes the MultiDisplayImeTests to no longer use the legacy input
injection methods, and instead use the new recommended UinputTouchScreen
approach. This also adds a couple of assertions and waiting for
conditions to further deflake these tests.

This also fixes an issue introduced in [1], namely during
testNoConfigurationChangedWhenSwitchBetweenTwoIdenticalDisplays, where
the test originally tapped on the center of the second display to focus
it, but was accidentally changed to tap on the first display.

  [1]: I0739e5d36a364091dd4f0e5d07027438e2450b5c

Flag: EXEMPT testfix
Bug: 425653443
Test: atest MultiDisplayImeTests MultiDisplaySecurityImeTests
Change-Id: I6e9066d9fdac8d394efecc688db4797f11a84da8