CTS Telecom tests - minor fix
Make a few minor changes to the telecom CTS tests to improve reliability.
Asserting the number of calls using `getInCallService` was racy
because the ICS may not have bound at that point.
`getLastAddedCall()` internally calls `waitOnInCallService`
which ensures the ICS is bound.  Changing the order so that we
get the call first ensures we've waited for binding which should
fix the flakes.

There was also another issue where we were assuming that the InCallService would still be bound at the end of the tests - that may not always be the case, if it crashes after the disconnect.  Adding null checking there to mitigate that failure.

Change-Id: I77d2ddfc30084206629e4b9a858fc48a45545f50
Bug: 431410741
Flag: NONE test only fix.
Test: This is a test fix.