Make logging more verbose and timeouts more liberal in TaskMoveTests
Tests are flaky and it seems like the root cause is a pretty strict
timeout for response to an intent to check whether task movability is
allowed on a display.

Let's increase this timeout from 1s to 5s and add more logs for easier
investigation in case it's not sufficient.

Bug: 432117911
Flag: com.android.window.flags.enable_window_repositioning_api
Test: atest CtsWindowManagerDeviceOther:TaskMoveTests
Change-Id: I9113c51bb3404858c588c4a32e05a89c6c37e28f