Fix SharedCameraTest failure
When the remote process is started during the setup, if the process already exists then test was failing. Based on the test failure rate and the logs, looks like the remote process is sometimes automatically restarted by the activity manager if it was killed and did not exit gracefully. Modified the setup to kill the process if it already exists, instead of failing the setup

Bug: 431828339
Test: Run CtsCameraTestCases
android.hardware.multiprocess.camera.cts.SharedCameraTest

Change-Id: I323a93409bb766fc381d7b084ea18a85bf576f67