media: fix alpha_size and renderable_type attrbs during eglChooseConfig
If EGL_ALPHA_SIZE is zero. color buffers with the smallest alpha
component size are preferred. Otherwise, color buffers with the largest
alpha component of at least the specified size are preferred. So if
required attributes are R8G8B8A0, then eglChooseConfig may choose
R10G10B10A2. This is not desired.

Also if the intent is to create EGLES 3.0 context then renderable type
can be used to indicate the same.

These are addressed.

Bug: 288875769
Test: atest CtsMediaCodecTestCases:DecodeEditEncodeTest
Flag: TEST_ONLY
Change-Id: I75891f88b993182437dba955d67d6ff073ed7af7