Fix flaky CtsVoiceRecognition test for Car
In automotive, for the full module execution of
CtsVoiceRecognitionTestCases, the setup stage of
RecognitionServiceMicIndicatorTest would often fail with
the error 'Can not find mic indicator'.

The isolated execution of this class usually passes, but full module
execution seems to fail due to side effects from other tests.

The test class already provided waitForNoIndicatorForCar, but this was
not being used in the setup. Using this method instead of
waitForNoIndicator gets rid of flakiness

Bug: 433543365
Change-Id: I8f785863e68b3807671edb4218a74783ec2e4bf9
Signed-off-by: Oleksandr Nikitin <oleksandr.nikitin@bmw.de>