Fix null pointer exceptions in ImeEventStream
If there is an assertion with an event name on an incorrect event
matcher (e.g. showSoftInput with editorMatcher), then the
ImeEventStream could throw a null pointer exception if that event is
found, as it won't have the required arguments (i.e. editorInfo). This
masks the underlying issue, and also stops the test execution early,
without showing the full ImeEventStream result.

To avoid these unexpected crashes, this modifies the matchers to check
if the required argument is null, in which case the predicate would
evaluate to false.

Flag: EXEMPT testfix
Bug: 420049339
Test: atest CtsInputMethodTestCases
Change-Id: Ie1a52e17c8f82de193b40f54950f3e5c5cbb30fc